Fix error when message file exists but UID has been deleted
When a user record has been removed from `/etc/passwd` but their message file
is still associated with that UID, Iris has no way of pulling up the
user information and crashes.

This corrects that issue, and skips the file.

Additionally, a "debug" mode was added to Iris, so that we can track
down these kinds of issues without annoying regular users of the
software.

Thanks go out to ubergeek for the bug report!

Issue report: https://github.com/Calamitous/iris/issues/4