use a hash set for unread topics
The "u" or "unread" command diffs two sets stored on disk: all topics, and unread messages.

A flat array was previously used for both, but this resulted in slow O(n^2) behavior.

This diff turns the unread messaegs into a hash set and uses that for the "contains" test in the loop over all topics.